+++ Sources/Moya/Endpoint.swift
@@ -14,7 +14,10 @@ public enum EndpointSampleResponse {
 }
 
 /// Class for reifying a target of the `Target` enum unto a concrete `Endpoint`.
-open class Endpoint<Target> {
+/// - Note: As of Moya 11.0.0 Endpoint is no longer generic.
+///   Existing code should work as is after removing the generic.
+///   See #1529 and #1524 for the discussion.
+open class Endpoint {
     public typealias SampleResponseClosure = () -> EndpointSampleResponse
 
     /// A string representation of the URL for the request.
@@ -46,12 +49,12 @@ open class Endpoint<Target> {
     }
 
     /// Convenience method for creating a new `Endpoint` with the same properties as the receiver, but with added HTTP header fields.
-    open func adding(newHTTPHeaderFields: [String: String]) -> Endpoint<Target> {
+    open func adding(newHTTPHeaderFields: [String: String]) -> Endpoint {
         return Endpoint(url: url, sampleResponseClosure: sampleResponseClosure, method: method, task: task, httpHeaderFields: add(httpHeaderFields: newHTTPHeaderFields))
     }
 
     /// Convenience method for creating a new `Endpoint` with the same properties as the receiver, but with replaced `task` parameter.
-    open func replacing(task: Task) -> Endpoint<Target> {
+    open func replacing(task: Task) -> Endpoint {
         return Endpoint(url: url, sampleResponseClosure: sampleResponseClosure, method: method, task: task, httpHeaderFields: httpHeaderFields)
     }
 
@@ -70,6 +73,7 @@ open class Endpoint<Target> {
 
 /// Extension for converting an `Endpoint` into a `URLRequest`.
 extension Endpoint {
+    // swiftlint:disable cyclomatic_complexity
     /// Returns the `Endpoint` converted to a `URLRequest` if valid. Throws an error otherwise.
     public func urlRequest() throws -> URLRequest {
         guard let requestURL = Foundation.URL(string: url) else {
@@ -88,6 +92,8 @@ extension Endpoint {
             return request
         case let .requestJSONEncodable(encodable):
             return try request.encoded(encodable: encodable)
+        case let .requestCustomJSONEncodable(encodable, encoder: encoder):
+            return try request.encoded(encodable: encodable, encoder: encoder)
         case let .requestParameters(parameters, parameterEncoding):
             return try request.encoded(parameters: parameters, parameterEncoding: parameterEncoding)
         case let .uploadCompositeMultipart(_, urlParameters):
@@ -108,6 +114,7 @@ extension Endpoint {
             return try bodyfulRequest.encoded(parameters: urlParameters, parameterEncoding: urlEncoding)
         }
     }
+    // swiftlint:enable cyclomatic_complexity
 }
 
 /// Required for using `Endpoint` as a key type in a `Dictionary`.
@@ -119,7 +126,7 @@ extension Endpoint: Equatable, Hashable {
 
     /// Note: If both Endpoints fail to produce a URLRequest the comparison will
     /// fall back to comparing each Endpoint's hashValue.
-    public static func == <T>(lhs: Endpoint<T>, rhs: Endpoint<T>) -> Bool {
+    public static func == (lhs: Endpoint, rhs: Endpoint) -> Bool {
         let lhsRequest = try? lhs.urlRequest()
         let rhsRequest = try? rhs.urlRequest()
         if lhsRequest != nil, rhsRequest == nil { return false }