+++ Sources/RxMoya/Observable+Response.swift
@@ -54,9 +54,9 @@ extension ObservableType where E == Response {
     }
 
     /// Maps received data at key path into a Decodable object. If the conversion fails, the signal errors.
-    public func map<D: Decodable>(_ type: D.Type, atKeyPath keyPath: String? = nil, using decoder: JSONDecoder = JSONDecoder()) -> Observable<D> {
+    public func map<D: Decodable>(_ type: D.Type, atKeyPath keyPath: String? = nil, using decoder: JSONDecoder = JSONDecoder(), failsOnEmptyData: Bool = true) -> Observable<D> {
         return flatMap { response -> Observable<D> in
-            return Observable.just(try response.map(type, atKeyPath: keyPath, using: decoder))
+            return Observable.just(try response.map(type, atKeyPath: keyPath, using: decoder, failsOnEmptyData: failsOnEmptyData))
         }
     }
 }